SETUP
Quantum Computing: Set up a cloud-based quantum computing service that provides access to a quantum computer capable of executing quantum algorithms. This can be done through platforms like IBM Quantum or Google Quantum Computing.
Data Analytics: Prepare the large dataset that needs to be processed. Ensure that the data is in a format compatible with the quantum algorithms and can be efficiently processed using MapReduce.
Virtual Reality: Set up a virtual reality environment using a framework like A-Frame. Create the necessary HTML files and include the required libraries to render the virtual reality scene.